DPS: Add option to disable communication with Trusted Party

+++ distributed_prov_system/provenance/apps.py
@@ -11,22 +11,23 @@ class ProvenanceConfig(AppConfig):
     name = 'provenance'
 
     def ready(self):
-        resp = requests.get(f'http://{config.tp_fqdn}/api/v1/info')
+        if not config.disable_tp:
+            resp = requests.get(f'http://{config.tp_fqdn}/api/v1/info')
 
-        assert resp.ok, "Couldn't retrieve info from TP!"
-        info = json.loads(resp.content)
+            assert resp.ok, "Couldn't retrieve info from TP!"
+            info = json.loads(resp.content)
 
-        try:
-            DefaultTrustedParty.nodes.get(identifier=info['id'])
-        except DoesNotExist:
-            # Only one DefaultTP expected!
-            for node in DefaultTrustedParty.nodes.all():
-                node.delete()
+            try:
+                DefaultTrustedParty.nodes.get(identifier=info['id'])
+            except DoesNotExist:
+                # Only one DefaultTP expected!
+                for node in DefaultTrustedParty.nodes.all():
+                    node.delete()
 
-            tp = DefaultTrustedParty()
-            tp.identifier = info['id']
-            tp.url = config.tp_fqdn
-            tp.certificate = info['certificate']
-            tp.valid = True
-            tp.checked = True
-            tp.save()
+                tp = DefaultTrustedParty()
+                tp.identifier = info['id']
+                tp.url = config.tp_fqdn
+                tp.certificate = info['certificate']
+                tp.valid = True
+                tp.checked = True
+                tp.save()
